Don't fail silently when grammar is missing
These conditional clauses _are_ found in some `-ts` modes, but only in the case where the parent mode is the non-`ts` version of the mode: in such cases, the clause quietly "downgrades" the mode to the non-`ts` version when the grammar is missing, so that the `-ts` mode can be enabled unconditionally in `auto-mode-alist`.

In other cases, like `noir-ts-mode`, the mode is non-functional without the grammar and should throw an error instead.

It is unclear how best to handle the `###autoload`-ed addition to `auto-mode-alist`, since we'd expect to be able to install the same package config in Emacsen with and without tree-sitter. It's probably safest to guard the `auto-mode-alist` expression in a `(when (treesit-read-p 'lang) ...)` form too.